% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movmean-package.R
\docType{package}
\name{movmean-package}
\alias{movmean}
\alias{movmean-package}
\title{movmean: What the Package Does (One Line, Title Case)}
\description{
moving average.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

Authors:
\itemize{
  \item CUGer \email{users@cug.edu.cn}
}

}
\keyword{internal}
