% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{movmean_rcpp}
\alias{movmean_rcpp}
\title{movmean}
\usage{
movmean_rcpp(x, halfwin = 1L, w = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{halfwin}{Integer, half of moving window size}

\item{w}{Corresponding weights of xx, same long as yy.}
}
\description{
NA and Inf values in the xx will be ignored automatically.
}
\examples{
x <- 1:100
x[50] <- NA; x[80] <- Inf
s1 <- movmean_rcpp(x, 1)
s2 <- movmean_rcpp(x, 2)
}
