% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/movmean-package.R,
%   R/movmean_julia.R, R/movmean_r.R
\name{movmean}
\alias{movmean}
\alias{movmean_rcpp}
\alias{movmean_julia}
\alias{movmean_r}
\title{moving average

NA and Inf values in the xx will be ignored automatically.}
\usage{
movmean_rcpp(x, halfwin = 1L, w = NULL)

movmean_julia(x, halfwin = 1L)

movmean_r(x, halfwin = 2L)
}
\arguments{
\item{x}{a numeric vector}

\item{halfwin}{integer, half size of the moving windows}

\item{w}{Corresponding weights of xx, same long as yy.}
}
\description{
moving average
}
\examples{
x <- 1:100
x[50] <- NA; x[80] <- Inf
s1 <- movmean_rcpp(x, 1)
s2 <- movmean_rcpp(x, 2)
x <- c(1, 2, 3, 4, 5)
r_r <- movmean_R(x, 2)
}
