% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movmean-package.R, R/movmean_juia.R
\name{movmean}
\alias{movmean}
\alias{movmean_julia}
\title{moving average}
\usage{
movmean_julia(x, halfwin = 1L)
}
\arguments{
\item{x}{a numeric vector}

\item{halfwin}{integer, half size of the moving windows}
}
\description{
moving average
}
