% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{movmean}
\alias{movmean}
\title{movmean}
\usage{
movmean(y, halfwin = 1L, w = NULL)
}
\arguments{
\item{y}{A numeric vector.}

\item{halfwin}{Integer, half of moving window size}

\item{w}{Corresponding weights of yy, same long as yy.}
}
\description{
NA and Inf values in the yy will be ignored automatically.
}
\examples{
x <- 1:100
x[50] <- NA; x[80] <- Inf
s1 <- movmean(x, 1)
s2 <- movmean(x, 2)
}
